Update index.html
Tuktak Oval Coin (TOC) is a TRC-20 token developed for blockchain education, simulation, and experimentation on the TRON network. This token is non-monetary and has no financial value or commercial use. 

It is intended to help developers, researchers, and students test token functionality in wallets, smart contracts, and dApps in a safe and realistic environment.

Our goal is to provide a learning tool for smart contract behavior, Web3 wallet interaction, token visibility, and token simulation across decentralized platforms.

This token is not listed on any exchange and should not be traded.

Official Website: https://tuktakoval.it.com  
Contact Email: contact@tuktakoval.it.com